refactor : BaseballGameCounts에서 BaseballGameCountsMessage 분리
